yandex-cloud-helper/pyproject.toml
[project]
name = "yandex-cloud-helper"
version = "0.0.1"
description = "Check README"
authors = []
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = ["inquirerpy (>=0.3.4,<0.4.0)", "python-dotenv-vault (>=0.6.4,<0.7.0)", "pygments (>=2.19.1,<3.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
ruff = "^0.9.1"
types-pygments = "^2.19.0.20250107"

[tool.mypy]
strict = true
exclude = [".venv"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
